add parsing for if stmt
different from an expression, if statement has only if, no else.
kind of bogus now, if expressions are split into compound and simple
expressions then the separate if stmt might not be necessary

also added a test to validate if stmt parsing and added comparison
for the Void value. Why are Nil and Void two separate things? One
is clearly a list, maybe that's enough?