disallow single expr if-only stmts
these were indistinguishable from first line of an
if-else block to the parser. which means it would be confusing
to humans as well, trying to figure out what they added
and would lead to bugs where people added to a single expr
if block thinking it was included in the block when it wasn't.

the if stmt block requires a closing -- so will give parse
errors to avoid this problem