structures can now be constructed
a constructor shows like a regular function. when executed,
it makes a structure of the expected type.
constructor func body has an op for each field in the structure.
maybe one op to do everything would be fine, but probably unnecessary
optimization right now.

also add a T test for the rgb structure script