fixed some borrow issues with Fiber/Frame
might not keep it, but there it is in the library now.
still trying to figure out what the interface to rust functions
should be. narrowed it down a bit now to:
* they get Fiber
* they get Frame and Handle

only difference there is slightly easier access to Frame and the
fiber_id is accessible from functions

in the olden days, the whole worker was accessible. should that
be allowed now?