finally got http serving working
still a lot of work to do, good start though.
rust handler does the code and returns a response.
need to fork the bind call or add a listen call or something
also need to remove all these ugly debug messages I put in
while trying to get it to work.