trying some things for semantic phase
was trying to interleave all the semantic phases,
but probably can't do that. At least not to the extent
that I was trying where the entire semantic pipeline
runs on each node before moving to the next node.

for example, doesn't make sense to assign registers, then delete nodes
later and have empty, unused registers