put register assignment back into code generation
want to be able to use the call stack allocation for register assignment
can add the RegTab / RegStack to the make_sub_ops call signature in
a way that I can't do w/ the walk interface for semantics.
Also semantic analysis and any optimizations are
definitely completely done by that point.