a few more steps towards fibers as tokio tasks
going to abort this now and shift back to leema Worker as
tokio tasks. Might eventually go back to this but want to
dial back the overall change of switching to async/await
to something that is closer to what previously was.

too many open questions about how to handle code loading
and io for individual fiber tasks and just how to use
async/await generally